replaced Double.pi with CGFloat.pi

+++ NVActivityIndicatorView/NVActivityIndicatorShape.swift
@@ -30,14 +30,14 @@ enum NVActivityIndicatorShape {
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
                 radius: size.width / 2,
                 startAngle: 0,
-                endAngle: CGFloat(2 * Double.pi),
+                endAngle: 2 * CGFloat.pi,
                 clockwise: false);
             layer.fillColor = color.cgColor
         case .circleSemi:
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
                 radius: size.width / 2,
-                startAngle: CGFloat(-Double.pi / 6),
-                endAngle: CGFloat(-5 * Double.pi / 6),
+                startAngle: -CGFloat.pi / 6,
+                endAngle: -5 * CGFloat.pi / 6,
                 clockwise: false)
             path.close()
             layer.fillColor = color.cgColor
@@ -45,52 +45,52 @@ enum NVActivityIndicatorShape {
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
                 radius: size.width / 2,
                 startAngle: 0,
-                endAngle: CGFloat(2 * Double.pi),
+                endAngle: 2 * CGFloat.pi,
                 clockwise: false);
             layer.fillColor = nil
             layer.strokeColor = color.cgColor
             layer.lineWidth = lineWidth
         case .ringTwoHalfVertical:
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
-                radius:size.width / 2,
-                startAngle:CGFloat(-3 * (Double.pi/4)),
-                endAngle:CGFloat(-(Double.pi/4)),
-                clockwise:true)
+                radius: size.width / 2,
+                startAngle: -3 * (CGFloat.pi/4),
+                endAngle: -(CGFloat.pi/4),
+                clockwise: true)
             path.move(
-                to: CGPoint(x: size.width / 2 - size.width / 2 * CGFloat(cos((Double.pi/4))),
-                    y: size.height / 2 + size.height / 2 * CGFloat(sin((Double.pi/4))))
+                to: CGPoint(x: size.width / 2 - size.width / 2 * cos(CGFloat.pi/4),
+                    y: size.height / 2 + size.height / 2 * sin(CGFloat.pi/4))
             )
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
-                radius:size.width / 2,
-                startAngle:CGFloat(-5 * (Double.pi/4)),
-                endAngle:CGFloat(-7 * (Double.pi/4)),
-                clockwise:false)
+                radius: size.width / 2,
+                startAngle: -5 * (CGFloat.pi/4),
+                endAngle: -7 * (CGFloat.pi/4),
+                clockwise: false)
             layer.fillColor = nil
             layer.strokeColor = color.cgColor
             layer.lineWidth = lineWidth
         case .ringTwoHalfHorizontal:
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
-                radius:size.width / 2,
-                startAngle:CGFloat(3 * (Double.pi/4)),
-                endAngle:CGFloat(5 * (Double.pi/4)),
-                clockwise:true)
+                radius: size.width / 2,
+                startAngle: 3 * (CGFloat.pi/4),
+                endAngle: 5 * (CGFloat.pi/4),
+                clockwise: true)
             path.move(
-                to: CGPoint(x: size.width / 2 + size.width / 2 * CGFloat(cos((Double.pi/4))),
-                    y: size.height / 2 - size.height / 2 * CGFloat(sin((Double.pi/4))))
+                to: CGPoint(x: size.width / 2 + size.width / 2 * cos(CGFloat.pi/4),
+                    y: size.height / 2 - size.height / 2 * sin(CGFloat.pi/4))
             )
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
-                radius:size.width / 2,
-                startAngle:CGFloat(-(Double.pi/4)),
-                endAngle:CGFloat((Double.pi/4)),
-                clockwise:true)
+                radius: size.width / 2,
+                startAngle: -(CGFloat.pi/4),
+                endAngle: (CGFloat.pi/4),
+                clockwise: true)
             layer.fillColor = nil
             layer.strokeColor = color.cgColor
             layer.lineWidth = lineWidth
         case .ringThirdFour:
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
                 radius: size.width / 2,
-                startAngle: CGFloat(-3 * (Double.pi/4)),
-                endAngle: CGFloat(-(Double.pi/4)),
+                startAngle: -3 * (CGFloat.pi/4),
+                endAngle: -(CGFloat.pi/4),
                 clockwise: false)
             layer.fillColor = nil
             layer.strokeColor = color.cgColor
@@ -117,7 +117,7 @@ enum NVActivityIndicatorShape {
             path.addArc(withCenter: CGPoint(x: size.width / 2, y: size.height / 2),
                 radius: size.width / 4,
                 startAngle: 0,
-                endAngle: CGFloat(2 * Double.pi),
+                endAngle: 2 * CGFloat.pi,
                 clockwise: true);
             layer.fillColor = nil
             layer.strokeColor = color.cgColor
